default template argument type is now int_fast32_t

+++ include/quantity/Quantity_base.h
@@ -4,9 +4,12 @@
 namespace Quantity_Base {
 using Fraction::frac;
 
-template <frac L = 0, frac M = 0, frac T = 0, frac I = 0, class Vtype = double>
+template <frac L = static_cast<int_fast32_t>(0),
+          frac M = static_cast<int_fast32_t>(0),
+          frac T = static_cast<int_fast32_t>(0),
+          frac I = static_cast<int_fast32_t>(0), class Vtype = double>
 class Quantity_base {
-public:
+ public:
   constexpr Quantity_base() = default;
   constexpr Quantity_base(double value) : value(value) {}
 
@@ -25,7 +28,7 @@ public:
   constexpr frac<> get_I() { return I; }
 
   template <frac LL, frac MM, frac TT, frac II>
-  constexpr bool is_same_dim_as(Quantity_base<LL, MM, TT, II>) {
+  constexpr bool is_same_dim_as(Quantity_base<LL, MM, TT, II> /*unused*/) {
     return (L == LL) && (M == MM) && (T == TT) && (I == II);
   }
 
@@ -41,8 +44,8 @@ public:
   }
 
   template <frac L2, frac M2, frac T2, frac I2>
-  constexpr Quantity_base<L + L2, M + M2, T + T2, I + I2>
-  mul(const Quantity_base<L2, M2, T2, I2> &x) {
+  constexpr Quantity_base<L + L2, M + M2, T + T2, I + I2> mul(
+      const Quantity_base<L2, M2, T2, I2> &x) {
     return {value * x.value};
   }
 
@@ -51,8 +54,8 @@ public:
   }
 
   template <frac L2, frac M2, frac T2, frac I2>
-  constexpr Quantity_base<L - L2, M - M2, T - T2, I - I2>
-  div(const Quantity_base<L2, M2, T2, I2> &x) {
+  constexpr Quantity_base<L - L2, M - M2, T - T2, I - I2> div(
+      const Quantity_base<L2, M2, T2, I2> &x) {
     return {value / x.value};
   }
 
@@ -64,7 +67,8 @@ public:
     return std::abs(value);
   }
 
-  template <auto N> constexpr Quantity_base<L * N, M * N, T * N, I * N> pow() {
+  template <auto N>
+  constexpr Quantity_base<L * N, M * N, T * N, I * N> pow() {
     return std::pow(value, N);
   }
 
@@ -82,49 +86,45 @@ public:
 
   // << operator
   template <frac LL, frac MM, frac TT, frac II>
-  constexpr friend std::ostream &
-  operator<<(std::ostream &os, const Quantity_base<LL, MM, TT, II> &frac);
+  constexpr friend std::ostream &operator<<(
+      std::ostream &os, const Quantity_base<LL, MM, TT, II> &frac);
 
   // + operator
   template <frac LL, frac MM, frac TT, frac II>
-  constexpr friend Quantity_base<LL, MM, TT, II>
-  operator+(Quantity_base<LL, MM, TT, II> x,
-            const Quantity_base<LL, MM, TT, II> &y);
+  constexpr friend Quantity_base<LL, MM, TT, II> operator+(
+      Quantity_base<LL, MM, TT, II> x, const Quantity_base<LL, MM, TT, II> &y);
 
   // - operator
   template <frac LL, frac MM, frac TT, frac II>
-  constexpr friend Quantity_base<LL, MM, TT, II>
-  operator-(Quantity_base<LL, MM, TT, II> x,
-            const Quantity_base<LL, MM, TT, II> &y);
+  constexpr friend Quantity_base<LL, MM, TT, II> operator-(
+      Quantity_base<LL, MM, TT, II> x, const Quantity_base<LL, MM, TT, II> &y);
   template <frac LL, frac MM, frac TT, frac II>
-  constexpr friend Quantity_base<LL, MM, TT, II>
-  operator-(Quantity_base<LL, MM, TT, II> x);
+  constexpr friend Quantity_base<LL, MM, TT, II> operator-(
+      Quantity_base<LL, MM, TT, II> x);
 
   // * operator
-  template <frac L1, frac M1, frac T1, frac I1, //
+  template <frac L1, frac M1, frac T1, frac I1,  //
             frac L2, frac M2, frac T2, frac I2>
-  constexpr friend Quantity_base<L1 + L2, M1 + M2, T1 + T2, I1 + I2>
-  operator*(Quantity_base<L1, M1, T1, I1> x,
-            const Quantity_base<L2, M2, T2, I2> &y);
+  constexpr friend Quantity_base<L1 + L2, M1 + M2, T1 + T2, I1 + I2> operator*(
+      Quantity_base<L1, M1, T1, I1> x, const Quantity_base<L2, M2, T2, I2> &y);
   template <frac L1, frac M1, frac T1, frac I1>
-  constexpr friend Quantity_base<L1, M1, T1, I1>
-  operator*(Quantity_base<L1, M1, T1, I1> x, const double &y);
+  constexpr friend Quantity_base<L1, M1, T1, I1> operator*(
+      Quantity_base<L1, M1, T1, I1> x, const double &y);
   template <frac L1, frac M1, frac T1, frac I1>
-  constexpr friend Quantity_base<L1, M1, T1, I1>
-  operator*(const double &y, Quantity_base<L1, M1, T1, I1> x);
+  constexpr friend Quantity_base<L1, M1, T1, I1> operator*(
+      const double &y, Quantity_base<L1, M1, T1, I1> x);
 
   // / operator
-  template <frac L1, frac M1, frac T1, frac I1, //
+  template <frac L1, frac M1, frac T1, frac I1,  //
             frac L2, frac M2, frac T2, frac I2>
-  constexpr friend Quantity_base<L1 - L2, M1 - M2, T1 - T2, I1 - I2>
-  operator/(Quantity_base<L1, M1, T1, I1> x,
-            const Quantity_base<L2, M2, T2, I2> &y);
+  constexpr friend Quantity_base<L1 - L2, M1 - M2, T1 - T2, I1 - I2> operator/(
+      Quantity_base<L1, M1, T1, I1> x, const Quantity_base<L2, M2, T2, I2> &y);
   template <frac L1, frac M1, frac T1, frac I1>
-  constexpr friend Quantity_base<L1, M1, T1, I1>
-  operator/(Quantity_base<L1, M1, T1, I1> x, const double &y);
+  constexpr friend Quantity_base<L1, M1, T1, I1> operator/(
+      Quantity_base<L1, M1, T1, I1> x, const double &y);
   template <frac L1, frac M1, frac T1, frac I1>
-  constexpr friend Quantity_base<-L1, -M1, -T1, -I1>
-  operator/(const double &y, Quantity_base<L1, M1, T1, I1> x);
+  constexpr friend Quantity_base<-L1, -M1, -T1, -I1> operator/(
+      const double &y, Quantity_base<L1, M1, T1, I1> x);
 
   // == operator
 };
@@ -132,21 +132,21 @@ public:
 template <frac L, frac M, frac T, frac I>
 constexpr std::ostream &operator<<(std::ostream &os,
                                    const Quantity_base<L, M, T, I> &quantity) {
-  os << quantity.value << " <" //
+  os << quantity.value << " <"  //
      << L.to_str() << "," << M.to_str() << "," << T.to_str() << ","
      << I.to_str() << ">";
   return os;
 }
 
 template <frac L, frac M, frac T, frac I>
-constexpr Quantity_base<L, M, T, I>
-operator+(Quantity_base<L, M, T, I> x, const Quantity_base<L, M, T, I> &y) {
+constexpr Quantity_base<L, M, T, I> operator+(
+    Quantity_base<L, M, T, I> x, const Quantity_base<L, M, T, I> &y) {
   return x.add(y);
 }
 
 template <frac L, frac M, frac T, frac I>
-constexpr Quantity_base<L, M, T, I>
-operator-(Quantity_base<L, M, T, I> x, const Quantity_base<L, M, T, I> &y) {
+constexpr Quantity_base<L, M, T, I> operator-(
+    Quantity_base<L, M, T, I> x, const Quantity_base<L, M, T, I> &y) {
   return x.sub(y);
 }
 template <frac L, frac M, frac T, frac I>
@@ -154,34 +154,32 @@ constexpr Quantity_base<L, M, T, I> operator-(Quantity_base<L, M, T, I> x) {
   return {-x.value};
 }
 
-template <frac L1, frac M1, frac T1, frac I1, //
+template <frac L1, frac M1, frac T1, frac I1,  //
           frac L2, frac M2, frac T2, frac I2>
-constexpr Quantity_base<L1 + L2, M1 + M2, T1 + T2, I1 + I2>
-operator*(Quantity_base<L1, M1, T1, I1> x,
-          const Quantity_base<L2, M2, T2, I2> &y) {
+constexpr Quantity_base<L1 + L2, M1 + M2, T1 + T2, I1 + I2> operator*(
+    Quantity_base<L1, M1, T1, I1> x, const Quantity_base<L2, M2, T2, I2> &y) {
   return x.mul(y);
 }
 template <frac L1, frac M1, frac T1, frac I1>
-constexpr Quantity_base<L1, M1, T1, I1>
-operator*(Quantity_base<L1, M1, T1, I1> x, const double &y) {
+constexpr Quantity_base<L1, M1, T1, I1> operator*(
+    Quantity_base<L1, M1, T1, I1> x, const double &y) {
   return {x.value * y};
 }
 template <frac L1, frac M1, frac T1, frac I1>
-constexpr Quantity_base<L1, M1, T1, I1>
-operator*(const double &y, Quantity_base<L1, M1, T1, I1> x) {
+constexpr Quantity_base<L1, M1, T1, I1> operator*(
+    const double &y, Quantity_base<L1, M1, T1, I1> x) {
   return {x.value * y};
 }
 
-template <frac L1, frac M1, frac T1, frac I1, //
+template <frac L1, frac M1, frac T1, frac I1,  //
           frac L2, frac M2, frac T2, frac I2>
-constexpr Quantity_base<L1 - L2, M1 - M2, T1 - T2, I1 - I2>
-operator/(Quantity_base<L1, M1, T1, I1> x,
-          const Quantity_base<L2, M2, T2, I2> &y) {
+constexpr Quantity_base<L1 - L2, M1 - M2, T1 - T2, I1 - I2> operator/(
+    Quantity_base<L1, M1, T1, I1> x, const Quantity_base<L2, M2, T2, I2> &y) {
   return x.div(y);
 }
 template <frac L1, frac M1, frac T1, frac I1>
-constexpr Quantity_base<L1, M1, T1, I1>
-operator/(Quantity_base<L1, M1, T1, I1> x, const double &y) {
+constexpr Quantity_base<L1, M1, T1, I1> operator/(
+    Quantity_base<L1, M1, T1, I1> x, const double &y) {
   return {x.value / y};
 }
 
@@ -230,4 +228,4 @@ constexpr Quantity_base<L, M, T, I> hypot(const Quantity_base<L, M, T, I> &x,
   return {std::hypot(x.value, y.value, z.value)};
 }
 
-} // namespace Quantity_Base
+}  // namespace Quantity_Base
